Don't destructively include pullquotes in the loop
When pullquotes are set to display in a waterfall, pullquotes were added regardless of the type of waterfall - which was causing them (and posts of the `post` post type) to be displayed on archive pages where they didn't belong.

This changeset ensures that pullquotes are only added to loops where A) `post_type` isn't specified or B) where the `post` post type appears in the `post_type` array.